SSCSCI-677: move compress/encode logic to utility class
